refactor(registration-react): Separate props-enhancer from connector in FormDistributionApi
